requirements.txt
google-cloud-aiplatform
PyPDF2
re
smtplib
streamlit